refactor: move data and languages to features directory
