--- src/_languages/index.ts
@@ -1,18 +1,0 @@
-/**
- * Data
- */
-export { TOOLS_LANGUAGES } from './data/tools.lng';
-
-/**
- * Layout
- */
-export { NAVBAR_LANGUAGES } from './layout/navbar.lng';
-export { FOOTER_LANGUAGES } from './layout/footer.lng';
-
-/**
- * Pages
- */
-export { HOME_LANGUAGES } from './pages/home.lng';
-// Tools
-export { PX_REM_LANGUAGES } from './pages/tools/px-rem.lng';
-export { COLORS_CONVERTER_LANGUAGES } from './pages/tools/colors-converter.lng';
