+++ src/app/tools/px-rem/page.tsx
@@ -4,12 +4,9 @@
 
 import { ArrowsHorizontal } from '@phosphor-icons/react';
 
-import { PxRemForm, InfoTable } from '@/features/px-rem';
+import { PxRemForm, InfoTable, PX_REM_CONVERSION, PX_REM_LANGUAGES } from '@/features/px-rem';
 
 import { useLanguage } from '@/ui/components/language/language-provider';
-
-import { PX_REM_CONVERSION } from '@/_data/px-rem';
-import { PX_REM_LANGUAGES } from '@/_languages';
 
 // #endregion
