+++ src/features/colors-converter/components/colors-form.tsx
@@ -10,14 +10,13 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import { useDebouncedCallback } from 'use-debounce';
 import { ZodError, z } from 'zod';
 
-import { useColorsConverter } from '@/features/colors-converter';
-
 import { useLanguage } from '@/ui/components/language/language-provider';
 import { Card } from '@/ui/components/ui/card';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/ui/components/ui/form';
 import { Input } from '@/ui/components/ui/input';
 
-import { COLORS_CONVERTER_LANGUAGES } from '@/_languages';
+import { useColorsConverter } from '../hooks/color-converter-hook';
+import { COLORS_CONVERTER_LANGUAGES } from '../languages/colors-converter.lng';
 
 // #endregion
