+++ src/features/colors-converter/index.ts
@@ -3,3 +3,6 @@ export { ColorsForm } from './components/colors-form';
 
 // Hooks
 export { ColorConverterProvider, useColorsConverter } from './hooks/color-converter-hook';
+
+// Languages
+export { COLORS_CONVERTER_LANGUAGES } from './languages/colors-converter.lng';
