+++ src/features/colors-converter/languages/colors-converter.lng.ts
