+++ src/features/home/index.ts
@@ -1,0 +1,2 @@
+// Languages
+export { HOME_LANGUAGES } from './languages/home.lng';
