+++ src/features/px-rem/index.ts
@@ -1,2 +1,9 @@
+// Components
 export { InfoTable } from './components/info-table';
 export { PxRemForm } from './components/px-rem-form';
+
+// Data
+export { PX_REM_CONVERSION } from './data/px-rem';
+
+// Languages
+export { PX_REM_LANGUAGES } from './languages/px-rem.lng';
