+++ src/ui/components/layout/languages/index.ts
@@ -1,0 +1,3 @@
+export { NAVBAR_LANGUAGES } from './modules/navbar.lng';
+export { FOOTER_LANGUAGES } from './modules/footer.lng';
+export { TOOLS_LANGUAGES } from './modules/tools.lng';
