+++ src/ui/components/layout/languages/modules/footer.lng.ts
