+++ src/ui/components/layout/languages/modules/navbar.lng.ts
