+++ src/ui/components/layout/languages/modules/tools.lng.ts
